// dist/components/itsk-pager/pager-type.ts
export type pagerType = {
    /**
     * Текущая страница
     */
    page?: number;

    /**
     * Количество записей на странице
     */
    pageSize?: number;
    /**
     * Количество страниц
     */
    count?: number;
    /**
     * Количество записей
     */
    totalCount?: number;
    /**
     * первая видимая страница
     */
    start?: number;
    /**
     * Последняя видимая страница
     */
    end?: number;
    /**
     * Список страниц
     */
    pages?: number[];

    /**
     * Инициализация из анонимного объекта
     */
}